pyproject.toml
[tool.poetry]
name = "pybr2022"
version = "0.1.0"
description = ""
authors = ["Marco Rougeth <rougeth@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.0.1"
httpx = "^0.23.0"
loguru = "^0.6.0"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
ipython = "^8.5.0"
ipdb = "^0.13.9"
pytest = "^7.1.3"
black = "^22.8.0"
isort = "^5.10.1"
pytest-asyncio = "^0.19.0"
coverage = "^6.4.4"
pytest-httpx = "^0.21.0"
pytest-cov = "^3.0.0"
factory-boy = "^3.2.1"

[tool.coverage.run]
branch = true
omit = [
    "tests/*",
]

[tool.coverage.report]
show_missing = true
precision = 2

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
